changes made to the data types of the gershgoring circles algorithm

+++ server/allTheMethods/src/main/java/com/example/allTheMethods/dto/request/LinearSystemsDataDto.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public record LinearSystemsDataDto(
-        int[][] matrix,
+        double[][] matrix,
         List<Optional<String>> variables
 ) {
 }
